version 0.1.10

+++ ewm-service/src/main/java/ru/practicum/explore/event/EventAdminController.java
@@ -17,9 +17,9 @@ public class EventAdminController {
 
     @GetMapping
     public List<EventFullDto> getEvents(
-            @RequestParam(required = false) List<Integer> users,
+            @RequestParam(required = false) List<Long> users,
             @RequestParam(required = false) List<String> states,
-            @RequestParam(required = false) List<Integer> categories,
+            @RequestParam(required = false) List<Long> categories,
             @RequestParam(required = false) String rangeStart,
             @RequestParam(required = false) String rangeEnd,
             @RequestParam(required = false, defaultValue = "0") Integer from,
